+++ activerecord/test/models/cpk/order.rb
@@ -5,6 +5,8 @@ module Cpk
     self.table_name = :cpk_orders
     self.primary_key = [:shop_id, :id]
 
+    alias_attribute :id_value, :id
+
     has_many :order_agreements, primary_key: :id
     has_many :books
   end